requirements.txt
Flask
numpy
pandas
scikit-learn
google-generativeai 
gunicorn
